edge-to-edge update
Set up `edge-to-edge` as a separate layout rather than a sub-option for `one column`

It is technically a "13 column" row for the sake of UCBLayout and LayoutBase as to differentiate form options between it and the `one column`

Edge-to-edge does not have the normal form options all the other layouts have but instead just receives defaults for said options. These defaults are passed into the template so everything is "filled out" properly.