Equal Column Height
Added a form option on sections for equal column height.
This option only shows up for 2, 3, and 4 column options.
All blocks within the column will be made equal height flex items.

In Layout Builder the "Add Block" button is unfortunately a flex item as well and so the preview differs slightly with the saved layout.