Column Refactors
Big refactors to how columns are built to be more dynamic rather than prewritten divs per column.
This decision was made to cleanup the code and to add in information for column eminence. Main and auxiliary columns need to be labeled as such for jump navs to render properly.

The column eminence mainly affects the two and three column layouts as they can have varying sizes based on user choice. Equal width columns (6-6 and 4-4-4) are treated all as "main"